+++ domain/src/main/java/com/sun/domain/config/RedisConfig.java
@@ -1,6 +1,7 @@
 package com.sun.domain.config;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
@@ -10,6 +11,7 @@ import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSeriali
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
+@ConditionalOnProperty(prefix = "spring.config.activate", name = "on-profile", havingValue = "db")
 public class RedisConfig {
 
     @Value("${spring.redis.host}")
